Migrate to ES6
