+++ src/public/assets/scripts/main.js
@@ -1,1 +1,1 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[0],{0:function(n,c,o){o("WfG0"),n.exports=o("zcrr")},"4VcE":function(n,c){},E8Ov:function(n,c){},WfG0:function(n,c,o){o("E8Ov"),o("4VcE")},zcrr:function(n,c){}},[[0,1]]]);
+!function(e){var n={};function t(i){if(n[i])return n[i].exports;var s=n[i]={i:i,l:!1,exports:{}};return e[i].call(s.exports,s,s.exports,t),s.l=!0,s.exports}t.m=e,t.c=n,t.d=function(e,n,i){t.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:i})},t.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},t.t=function(e,n){if(1&n&&(e=t(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var i=Object.create(null);if(t.r(i),Object.defineProperty(i,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var s in e)t.d(i,s,function(n){return e[n]}.bind(null,s));return i},t.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return t.d(n,"a",n),n},t.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},t.p="/",t(t.s=0)}({0:function(e,n,t){t("WfG0"),e.exports=t("zcrr")},"1YLx":function(e,n,t){"use strict";t.r(n),t("E8Ov").AnnotationsTabPlugins.assistanceRequest={props:{annotations:{type:Array,required:!0},url:{type:String,required:!0}},computed:{selectedAnnotations:function(){return this.annotations.filter((function(e){return e.selected}))},isDisabled:function(){return 1!==this.selectedAnnotations.length},href:function(){return this.url+this.selectedAnnotations[0].id}}}},"4VcE":function(e,n){},E8Ov:function(e,n,t){"use strict";t.r(n),t.d(n,"AnnotationCanvasComponent",(function(){return i})),t.d(n,"AnnotationsApi",(function(){return s})),t.d(n,"AnnotationsTabPlugins",(function(){return r})),t.d(n,"ImagesStore",(function(){return a})),t.d(n,"LabelTrees",(function(){return o})),t.d(n,"LoaderMixin",(function(){return u})),t.d(n,"Messages",(function(){return l})),t.d(n,"SidebarComponent",(function(){return c})),t.d(n,"SidebarTabComponent",(function(){return d})),t.d(n,"Typeahead",(function(){return h})),t.d(n,"UsersApi",(function(){return f}));var i=biigle.$require("annotations.components.annotationCanvas"),s=biigle.$require("api.annotations"),r=biigle.$require("annotations.components.annotationsTabPlugins"),a=biigle.$require("annotations.stores.images"),o=biigle.$require("labelTrees.components.labelTrees"),u=biigle.$require("core.mixins.loader"),l=biigle.$require("messages.store"),c=biigle.$require("core.components.sidebar"),d=biigle.$require("core.components.sidebarTab"),h=biigle.$require("core.components.typeahead"),f=biigle.$require("api.users")},N8jO:function(e,n,t){"use strict";t.r(n);var i=t("E8Ov"),s=new Vue({components:{labelTrees:i.LabelTrees,typeahead:i.Typeahead},data:{typeaheadTemplate:'<span v-text="item.name"></span><br><small v-text="item.affiliation"></small>',users:[],selectedUser:null},computed:{flatLabels:function(){var e=[];return this.labelTrees.forEach((function(n){Array.prototype.push.apply(e,n.labels)})),e},selectedLabels:function(){return this.flatLabels.filter((function(e){return e.selected}))},hasTooManySelectedLabels:function(){return this.selectedLabels.length>5},receiverId:function(){return this.selectedUser?this.selectedUser.id:""},receiverName:function(){return this.selectedUser?this.selectedUser.name:""},labelTrees:function(){return biigle.$require("ananas.labelTrees")}},methods:{close:function(){window.close()},loadUsers:function(){return i.UsersApi.query().then(this.usersLoaded,i.Messages.handleErrorResponse)},usersLoaded:function(e){e.data.forEach((function(e){e.name=e.firstname+" "+e.lastname})),Vue.set(this,"users",e.data)},selectUser:function(e){this.selectedUser=e},clearSelectedUser:function(){this.selectedUser=null}},created:function(){var e=this,n=biigle.$require("ananas.oldLabels");if(Array.isArray(n)){var t={};n.forEach((function(e){t[e]=null})),this.flatLabels.forEach((function(e){t.hasOwnProperty(e.id)&&Vue.set(e,"selected",!0)}))}var i=this.loadUsers(),s=biigle.$require("ananas.oldReceiverId");s&&i.then((function(){for(var n=e.users.length-1;n>=0;n--)if(e.users[n].id==s)return void(e.selectedUser=e.users[n])}))}}),r=Vue.resource("api/v1/annotation-assistance-requests{/id}",{},{respond:{method:"PUT",url:"api/v1/annotation-assistance-requests{/token}"}}),a={mixins:[i.LoaderMixin],components:{sidebar:i.SidebarComponent,sidebarTab:i.SidebarTabComponent,annotationCanvas:i.AnnotationCanvasComponent},data:{image:null,annotations:[]},computed:{annotation:function(){return biigle.$require("ananas.annotation")}},methods:{setImageAndAnnotation:function(e){this.image=e,this.annotations.push(this.annotation)},focusAnnotation:function(){this.$refs.canvas.focusAnnotation(this.annotation,!0)},handleLoadingError:function(e){i.Messages.danger(e)}},mounted:function(){this.startLoading(),i.ImagesStore.fetchAndDrawImage(this.annotation.image_id).then(this.setImageAndAnnotation).then(this.focusAnnotation).catch(this.handleLoadingError).finally(this.finishLoading)}},o=new Vue({mixins:[a],data:{pickedLabel:null,responseText:"",closed:!1,errors:[],showMinimap:!0},computed:{hasPickedLabel:function(){return null!==this.pickedLabel},hasResponseText:function(){return!!this.responseText},hasDisabledControls:function(){return this.closed||this.loading},hasErrors:function(){return this.errors.length>0},token:function(){return biigle.$require("ananas.token")}},methods:{pickLabel:function(e){this.hasDisabledControls||(this.pickedLabel===e?this.pickedLabel=null:this.pickedLabel=e)},submit:function(){var e={};this.hasPickedLabel&&(e.response_label_id=this.pickedLabel),this.hasResponseText&&(e.response_text=this.responseText),this.startLoading(),r.respond({token:token},e).finally(this.clearErrors).then(this.handleSuccess,this.handleError).finally(this.finishLoading)},handleSuccess:function(){this.closed=!0},clearErrors:function(e){return Vue.set(this,"errors",[]),e},handleError:function(e){for(var n in e.body)e.body.hasOwnProperty(n)&&this.errors.push(e.body[n][0])},checkForMobile:function(){window.innerWidth<1e3?this.showMinimap=!1:this.showMinimap=!0}},mounted:function(){this.checkForMobile(),window.addEventListener("resize",this.checkForMobile)}}),u=new Vue({mixins:[a],computed:{attachedSuggestedLabel:function(){for(var e=this.existingLabels,n=e.length-1;n>=0;n--)if(e[n].label_id===this.suggestedLabelId&&e[n].user_id===this.userId)return!0;return!1},existingLabels:function(){return biigle.$require("ananas.labels")},annotation:function(){return biigle.$require("ananas.annotation")},userId:function(){return biigle.$require("ananas.userId")},suggestedLabelId:function(){return biigle.$require("ananas.suggestedLabelId")}},methods:{attach:function(){this.startLoading(),i.AnnotationsApi.attachLabel({id:this.annotation.id},{label_id:this.suggestedLabelId,confidence:1}).then(this.handleAttachSuccess,i.Messages.handleErrorResponse).finally(this.finishLoading)},handleAttachSuccess:function(){this.existingLabels.push({label_id:this.suggestedLabelId,user_id:this.userId})}}});biigle.$mount("ananas-respond-container",o),biigle.$mount("ananas-show-container",u),biigle.$mount("create-ananas-form",s)},WfG0:function(e,n,t){t("E8Ov"),t("4VcE"),t("1YLx"),t("N8jO")},zcrr:function(e,n){}});
