+++ src/public/assets/scripts/vendor.js
@@ -1,0 +1,1 @@
+(window["webpackJsonp"] = window["webpackJsonp"] || []).push([["/assets/scripts/vendor"],[]]);
