+++ src/public/mix-manifest.json
@@ -1,5 +1,4 @@
 {
-    "/assets/scripts/main.js": "/assets/scripts/main.js?id=c48ddb9ebe8017912c7e",
-    "/assets/styles/main.css": "/assets/styles/main.css?id=48349a23ba30aece45cd",
-    "/assets/scripts/manifest.js": "/assets/scripts/manifest.js?id=3c768977c2574a34506e"
+    "/assets/scripts/main.js": "/assets/scripts/main.js?id=f9688cd967fbd1fb7afe",
+    "/assets/styles/main.css": "/assets/styles/main.css?id=48349a23ba30aece45cd"
 }
