+++ src/resources/assets/js/main.js
@@ -1,3 +1,4 @@
 require('./import');
 require('./export');
-// TODO
+require('./components/annotationsTabPlugin');
+require('./mounts');
