+++ src/resources/assets/js/mounts.js
@@ -1,0 +1,8 @@
+import CreateForm from './createForm';
+import RespondContainer from './respondContainer';
+import ShowContainer from './showContainer';
+
+biigle.$mount('ananas-respond-container', RespondContainer);
+biigle.$mount('ananas-show-container', ShowContainer);
+biigle.$mount('create-ananas-form', CreateForm);
+
