+++ src/resources/views/settings/notifications.blade.php
@@ -22,37 +22,34 @@
 
 @push('scripts')
 <script type="text/javascript">
-    biigle.$viewModel('ananas-notification-settings', function (element) {
-        new Vue({
-            el: element,
-            mixins: [biigle.$require('core.mixins.loader')],
-            data: {
-                settings: '{!! $user->getSettings('ananas_notifications', config('ananas.notifications.default_settings')) !!}',
-                saved: false,
-                error: false,
+    biigle.$mount('ananas-notification-settings', new Vue({
+        mixins: [biigle.$require('core.mixins.loader')],
+        data: {
+            settings: '{!! $user->getSettings('ananas_notifications', config('ananas.notifications.default_settings')) !!}',
+            saved: false,
+            error: false,
+        },
+        methods: {
+            handleSuccess: function () {
+                this.saved = true;
+                this.error = false;
             },
-            methods: {
-                handleSuccess: function () {
-                    this.saved = true;
-                    this.error = false;
-                },
-                handleError: function (response) {
-                    this.saved = false;
-                    this.error = true;
-                    biigle.$require('messages.store').handleErrorResponse(response);
-                },
+            handleError: function (response) {
+                this.saved = false;
+                this.error = true;
+                biigle.$require('messages.store').handleErrorResponse(response);
             },
-            watch: {
-                settings: function (settings) {
-                    this.startLoading();
-                    this.$http.post('api/v1/users/my/settings/ananas', {
-                            ananas_notifications: this.settings,
-                        })
-                        .then(this.handleSuccess, this.handleError)
-                        .finally(this.finishLoading);
-                },
+        },
+        watch: {
+            settings: function (settings) {
+                this.startLoading();
+                this.$http.post('api/v1/users/my/settings/ananas', {
+                        ananas_notifications: this.settings,
+                    })
+                    .then(this.handleSuccess, this.handleError)
+                    .finally(this.finishLoading);
             },
-        });
-    });
+        },
+    }));
 </script>
 @endpush
